+++ core/src/io/anuke/mindustry/world/blocks/logic/MemoryBlock.java
@@ -9,7 +9,7 @@ public class MemoryBlock extends LogicBlock{
     }
 
     @Override
-    public byte signal(Tile tile){
+    public int signal(Tile tile){
         return tile.<LogicEntity>entity().signal;
     }
 }
